+++ src/clojure_lsp/feature/completion.clj
@@ -203,8 +203,7 @@
          (map #(element->completion-item % nil :refer)))))
 
 (defn ^:private with-elements-from-alias [cursor-loc cursor-alias cursor-value matches-fn db]
-  (when-let [aliases (some->> (:analysis @db)
-                              (q/filter-project-analysis)
+  (when-let [aliases (some->> (q/filter-project-analysis (:analysis @db) db)
                               (mapcat val)
                               (filter #(identical? :namespace-alias (:bucket %))))]
     (let [alias-namespaces (->> aliases
@@ -309,11 +308,9 @@
         analysis (:analysis @db)
         current-ns-elements (get analysis filename)
         support-snippets? (get-in @db [:client-capabilities :text-document :completion :completion-item :snippet-support] false)
-        other-ns-elements (->> (dissoc analysis filename)
-                               q/filter-project-analysis
+        other-ns-elements (->> (q/filter-project-analysis (dissoc analysis filename) db)
                                (mapcat val))
-        external-ns-elements (->> (dissoc analysis filename)
-                                  q/filter-external-analysis
+        external-ns-elements (->> (q/filter-external-analysis (dissoc analysis filename) db)
                                   (mapcat val))
         cursor-loc (when-let [loc (parser/safe-loc-at-pos text row col)]
                      (when (or (not (-> loc z/node meta))
@@ -393,7 +390,7 @@
         definition (q/find-definition analysis {:filename filename
                                                 :name (symbol name)
                                                 :to (symbol ns)
-                                                :bucket :var-usages})]
+                                                :bucket :var-usages} db)]
     (if definition
       (-> item
           (assoc :documentation (f.hover/hover-documentation definition db))