+++ test/clojure_lsp/queries_test.clj
@@ -15,13 +15,13 @@
     (h/load-code-and-locs "(ns foo.bar)" (h/file-uri "file:///a.clj"))
     (h/load-code-and-locs "(ns foo.bar)" (h/file-uri "file:///b.clj"))
     (h/load-code-and-locs "(ns foo.bar)" (h/file-uri "jar:file:///some.jar!/some-file.clj"))
-    (is (= 2 (count (q/filter-project-analysis (:analysis @db/db))))))
+    (is (= 2 (count (q/filter-project-analysis (:analysis @db/db) db/db)))))
   (testing "when dependency-scheme is jar"
     (swap! db/db medley/deep-merge {:settings {:dependency-scheme "jar"}})
     (h/load-code-and-locs "(ns foo.bar)" (h/file-uri "file:///a.clj"))
     (h/load-code-and-locs "(ns foo.bar)" (h/file-uri "file:///b.clj"))
     (h/load-code-and-locs "(ns foo.bar)" (h/file-uri "jar:file:///some.jar!/some-file.clj"))
-    (is (= 2 (count (q/filter-project-analysis (:analysis @db/db)))))))
+    (is (= 2 (count (q/filter-project-analysis (:analysis @db/db) db/db))))))
 
 (deftest external-analysis
   (testing "when dependency-scheme is zip"
@@ -29,27 +29,27 @@
     (h/load-code-and-locs "(ns foo.bar)" (h/file-uri "file:///a.clj"))
     (h/load-code-and-locs "(ns foo.bar)" (h/file-uri "file:///b.clj"))
     (h/load-code-and-locs "(ns foo.bar)" (h/file-uri "jar:file:///some.jar!/some-file.clj"))
-    (is (= 1 (count (q/filter-external-analysis (:analysis @db/db))))))
+    (is (= 1 (count (q/filter-external-analysis (:analysis @db/db) db/db)))))
   (testing "when dependency-scheme is jar"
     (swap! db/db medley/deep-merge {:settings {:dependency-scheme "jar"}})
     (h/load-code-and-locs "(ns foo.bar)" (h/file-uri "file:///a.clj"))
     (h/load-code-and-locs "(ns foo.bar)" (h/file-uri "file:///b.clj"))
     (h/load-code-and-locs "(ns foo.bar)" (h/file-uri "jar:file:///some.jar!/some-file.clj"))
-    (is (= 1 (count (q/filter-external-analysis (:analysis @db/db)))))))
+    (is (= 1 (count (q/filter-external-analysis (:analysis @db/db) db/db))))))
 
 (deftest find-last-order-by-project-analysis
   (testing "with pred that applies for both project and external analysis"
     (h/clean-db!)
     (h/load-code-and-locs "(ns foo.bar)" (h/file-uri "jar:file:///some.jar!/some-file.clj"))
     (h/load-code-and-locs "(ns foo.bar)" (h/file-uri "file:///a.clj"))
-    (let [element (#'q/find-last-order-by-project-analysis #(= 'foo.bar (:name %)) (:analysis @db/db))]
+    (let [element (#'q/find-last-order-by-project-analysis #(= 'foo.bar (:name %)) (:analysis @db/db) db/db)]
       (is (= (h/file-path "/a.clj") (:filename element)))))
   (testing "with pred that applies for both project and external analysis with multiple on project"
     (h/clean-db!)
     (h/load-code-and-locs "(ns foo.bar)" (h/file-uri "jar:file:///some.jar!/some-file.clj"))
     (h/load-code-and-locs "(ns foo.bar)" (h/file-uri "file:///a.clj"))
     (h/load-code-and-locs "(ns foo.bar)" (h/file-uri "file:///b.clj"))
-    (let [element (#'q/find-last-order-by-project-analysis #(= 'foo.bar (:name %)) (:analysis @db/db))]
+    (let [element (#'q/find-last-order-by-project-analysis #(= 'foo.bar (:name %)) (:analysis @db/db) db/db)]
       (is (= (h/file-path "/b.clj") (:filename element))))))
 
 (deftest find-element-under-cursor
@@ -96,24 +96,24 @@
     (h/assert-submaps
       [{:name 'x :name-row x-r :name-col x-c}
        {:name 'x :name-row x-use-r :name-col x-use-c}]
-      (q/find-references-from-cursor ana (h/file-path "/a.clj") x-r x-c true))
+      (q/find-references-from-cursor ana (h/file-path "/a.clj") x-r x-c true db/db))
     (h/assert-submaps
       [{:name 'filename :name-row param-r :name-col param-c}
        {:name 'filename :name-row param-use-r :name-col param-use-c}]
-      (q/find-references-from-cursor ana (h/file-path "/a.clj") param-r param-c true))
+      (q/find-references-from-cursor ana (h/file-path "/a.clj") param-r param-c true db/db))
     (h/assert-submaps
       ['{:name unknown}]
-      (q/find-references-from-cursor ana (h/file-path "/a.clj") unknown-r unknown-c true))
+      (q/find-references-from-cursor ana (h/file-path "/a.clj") unknown-r unknown-c true db/db))
     (h/assert-submaps
       [{:alias 'f-alias :name-row alias-r :name-col alias-c}
        {:alias 'f-alias :name 'foo :name-row alias-use-r :name-col alias-use-c}]
-      (q/find-references-from-cursor ana (h/file-path "/a.clj") alias-r alias-c true))
+      (q/find-references-from-cursor ana (h/file-path "/a.clj") alias-r alias-c true db/db))
     (h/assert-submaps
       [{:name "foo-kw" :name-row a-foo-kw-r :name-col a-foo-kw-c}
        {:name "foo-kw" :name-row b-foo-kw-r :name-col b-foo-kw-c}
        {:name "foo-kw" :name-row d-foo-kw-r :name-col d-foo-kw-c}
        {:name "foo-kw" :name-row c-foo-kw-r :name-col c-foo-kw-c}]
-      (q/find-references-from-cursor ana (h/file-path "/c.clj") b-foo-kw-r b-foo-kw-c true))))
+      (q/find-references-from-cursor ana (h/file-path "/c.clj") b-foo-kw-r b-foo-kw-c true db/db))))
 
 (deftest find-references-from-cursor-cljc
   (let [code (str "(ns a.b.c (:require [d.e.f :as |f-alias]))\n"
@@ -130,18 +130,18 @@
     (h/assert-submaps
       [{:name 'x :name-row x-r :name-col x-c}
        {:name 'x :name-row x-use-r :name-col x-use-c}]
-      (q/find-references-from-cursor ana (h/file-path "/a.cljc") x-r x-c true))
+      (q/find-references-from-cursor ana (h/file-path "/a.cljc") x-r x-c true db/db))
     (h/assert-submaps
       [{:name 'filename :name-row param-r :name-col param-c}
        {:name 'filename :name-row param-use-r :name-col param-use-c}]
-      (q/find-references-from-cursor ana (h/file-path "/a.cljc") param-r param-c true))
+      (q/find-references-from-cursor ana (h/file-path "/a.cljc") param-r param-c true db/db))
     (h/assert-submaps
       ['{:name unknown}]
-      (q/find-references-from-cursor ana (h/file-path "/a.cljc") unknown-r unknown-c true))
+      (q/find-references-from-cursor ana (h/file-path "/a.cljc") unknown-r unknown-c true db/db))
     (h/assert-submaps
       [{:alias 'f-alias :name-row alias-r :name-col alias-c}
        {:alias 'f-alias :name 'foo :name-row alias-use-r :name-col alias-use-c}]
-      (q/find-references-from-cursor ana (h/file-path "/a.cljc") alias-r alias-c true))))
+      (q/find-references-from-cursor ana (h/file-path "/a.cljc") alias-r alias-c true db/db))))
 
 (deftest find-references-from-defrecord
   (let [code (str "(defrecord |MyRecord [])\n"
@@ -157,7 +157,7 @@
       [{:name 'MyRecord :bucket :var-usages :name-row raw-r :name-col raw-c}
        {:name '->MyRecord :bucket :var-usages :name-row to-r :name-col to-c}
        {:name 'map->MyRecord :bucket :var-usages :name-row map-to-r :name-col map-to-c}]
-      (q/find-references-from-cursor ana (h/file-path "/a.clj") def-r def-c false))))
+      (q/find-references-from-cursor ana (h/file-path "/a.clj") def-r def-c false db/db))))
 
 (deftest find-references-excluding-function-different-arity
   (let [a-code (h/code "(ns a)"
@@ -187,7 +187,7 @@
            :from b
            :name-col 4
            :from-var bar}]
-        (q/find-references-from-cursor ana (h/file-path "/a.clj") bar-def-r bar-def-c false)))
+        (q/find-references-from-cursor ana (h/file-path "/a.clj") bar-def-r bar-def-c false db/db)))
     (testing "from usage"
       (h/assert-submaps
         '[{:name-row 6
@@ -201,7 +201,7 @@
            :from b
            :name-col 4
            :from-var bar}]
-        (q/find-references-from-cursor ana (h/file-path "/a.clj") bar-usa-r bar-usa-c false)))
+        (q/find-references-from-cursor ana (h/file-path "/a.clj") bar-usa-r bar-usa-c false db/db)))
     (testing "from other ns"
       (h/assert-submaps
         '[{:name-row 6
@@ -215,7 +215,7 @@
            :from b
            :name-col 4
            :from-var bar}]
-        (q/find-references-from-cursor ana (h/file-path "/b.clj") bar-usa-b-r bar-usa-b-c false)))))
+        (q/find-references-from-cursor ana (h/file-path "/b.clj") bar-usa-b-r bar-usa-b-c false db/db)))))
 
 (deftest find-definition-from-cursor
   (let [code (str "(ns a.b.c (:require [d.e.f :as |f-alias]))\n"
@@ -232,18 +232,18 @@
         ana (:analysis @db/db)]
     (h/assert-submap
       {:name 'x :name-row x-r :name-col x-c}
-      (q/find-definition-from-cursor ana (h/file-path "/a.clj") x-use-r x-use-c))
+      (q/find-definition-from-cursor ana (h/file-path "/a.clj") x-use-r x-use-c db/db))
     (h/assert-submap
       {:name 'filename :name-row param-r :name-col param-c}
-      (q/find-definition-from-cursor ana (h/file-path "/a.clj") param-use-r param-use-c))
+      (q/find-definition-from-cursor ana (h/file-path "/a.clj") param-use-r param-use-c db/db))
     (is (= nil
-           (q/find-definition-from-cursor ana (h/file-path "/a.clj") unknown-r unknown-c)))
+           (q/find-definition-from-cursor ana (h/file-path "/a.clj") unknown-r unknown-c db/db)))
     (h/assert-submap
       {:name 'foo :filename (h/file-path "/b.clj") :ns 'd.e.f}
-      (q/find-definition-from-cursor ana (h/file-path "/a.clj") alias-use-r alias-use-c))
+      (q/find-definition-from-cursor ana (h/file-path "/a.clj") alias-use-r alias-use-c db/db))
     (h/assert-submap
       {:alias 'f-alias :name-row alias-r :name-col alias-c}
-      (q/find-definition-from-cursor ana (h/file-path "/a.clj") alias-r alias-c))))
+      (q/find-definition-from-cursor ana (h/file-path "/a.clj") alias-r alias-c db/db))))
 
 (deftest find-definition-from-cursor-when-duplicate-from-external-analysis
   (let [_ (h/load-code-and-locs (h/code "(ns foo) (def bar)") "jar:file:///some.jar!/some-jar.clj")
@@ -253,7 +253,7 @@
         ana (:analysis @db/db)]
     (h/assert-submap
       {:name 'bar :filename (h/file-path "/a.clj")}
-      (q/find-definition-from-cursor ana (h/file-path "/b.clj") bar-r bar-c))))
+      (q/find-definition-from-cursor ana (h/file-path "/b.clj") bar-r bar-c db/db))))
 
 (deftest find-definition-from-cursor-when-it-has-same-namespace-from-clj-and-cljs
   (h/load-code-and-locs (h/code "(ns foo) (def bar)") (h/file-uri "jar:file:///some.jar!/some-jar.clj"))
@@ -264,21 +264,21 @@
           ana (:analysis @db/db)]
       (h/assert-submap
         {:name 'bar :filename (h/file-path "/some.jar:some-jar.clj")}
-        (q/find-definition-from-cursor ana (h/file-path "/b.clj") bar-r bar-c))))
+        (q/find-definition-from-cursor ana (h/file-path "/b.clj") bar-r bar-c db/db))))
   (testing "when on a cljs file"
     (let [[[bar-r bar-c]] (h/load-code-and-locs (h/code "(ns baz (:require [foo :as f]))"
                                                         "|f/bar") (h/file-uri "file:///b.cljs"))
           ana (:analysis @db/db)]
       (h/assert-submap
         {:name 'bar :filename (h/file-path "/some.jar:other-jar.cljs")}
-        (q/find-definition-from-cursor ana (h/file-path "/b.cljs") bar-r bar-c))))
+        (q/find-definition-from-cursor ana (h/file-path "/b.cljs") bar-r bar-c db/db))))
   (testing "when on a cljc file"
     (let [[[bar-r bar-c]] (h/load-code-and-locs (h/code "(ns baz (:require [foo :as f]))"
                                                         "|f/bar") (h/file-uri "file:///b.cljc"))
           ana (:analysis @db/db)]
       (h/assert-submap
         {:name 'bar :filename (h/file-path "/some.jar:some-jar.clj")}
-        (q/find-definition-from-cursor ana (h/file-path "/b.cljc") bar-r bar-c))))
+        (q/find-definition-from-cursor ana (h/file-path "/b.cljc") bar-r bar-c db/db))))
   (testing "when on a cljc file with multiple langs available"
     (let [[[bar-r-clj bar-c-clj]
            [bar-r-cljs bar-c-cljs]] (h/load-code-and-locs (h/code "(ns baz #?(:clj (:require [foo :as fc]) :cljs (:require [foo :as fs])))"
@@ -287,10 +287,10 @@
           ana (:analysis @db/db)]
       (h/assert-submap
         {:name 'bar :filename (h/file-path "/some.jar:some-jar.clj")}
-        (q/find-definition-from-cursor ana (h/file-path "/b.cljc") bar-r-clj bar-c-clj))
+        (q/find-definition-from-cursor ana (h/file-path "/b.cljc") bar-r-clj bar-c-clj db/db))
       (h/assert-submap
         {:name 'bar :filename (h/file-path "/some.jar:other-jar.cljs")}
-        (q/find-definition-from-cursor ana (h/file-path "/b.cljc") bar-r-cljs bar-c-cljs)))))
+        (q/find-definition-from-cursor ana (h/file-path "/b.cljc") bar-r-cljs bar-c-cljs db/db)))))
 
 (deftest find-definition-from-cursor-when-declared
   (let [[[bar-r bar-c]] (h/load-code-and-locs
@@ -301,7 +301,7 @@
         ana (:analysis @db/db)]
     (h/assert-submap
       {:name 'bar :filename (h/file-path "/a.clj") :defined-by 'clojure.core/defn :row 4}
-      (q/find-definition-from-cursor ana (h/file-path "/a.clj") bar-r bar-c))))
+      (q/find-definition-from-cursor ana (h/file-path "/a.clj") bar-r bar-c db/db))))
 
 (deftest find-unused-aliases
   (testing "clj"